requirements.txt
attrs
jupyter
matplotlib
pandas
parmap
scipy
scikit-learn


